requirements.txt
numpy
scipy
pillow
tensorflow
